ras: Initial bitmap font rendering
Material index used to identify the char data for now. A material
atlas should instead manage the bitmap assets in a subsequent change.

+++ rasgl/src/core/include/rasgl/core/text.h
@@ -22,7 +22,7 @@ RasResult core_draw_text(
     RenderState* state,
     RasFont* font,
     const char* text,
-    Point2i pos,
+    Point2f pos,
     uint8_t fg_color,
     int32_t bg_color);
